Step 1: Controller 実装（コメントの下に実装する）

+++ app/Http/Controllers/Api/ProjectController.php
@@ -30,7 +30,11 @@ class ProjectController extends ApiController
      */
     public function store(Request $request): JsonResponse
     {
-        # バリデーションを記載
+        # バリデーションを記載【Lesson2】
+        $validated = $request->validate([
+            'name' => 'required',
+            'is_archived' => 'required',
+        ]);
 
         // プロジェクト作成
         $project = Project::create([
@@ -38,8 +42,9 @@ class ProjectController extends ApiController
             'is_archived' => $request->is_archived ?? false,
         ]);
 
-        // 作成者を自動的にオーナーとして追加（users()リレーションのattach()を使用）
-
+        // 作成者を自動的にオーナーとして追加（users()リレーションのattach()を使用）【Lesson2】
+        $user = $request->user(); 
+        $project->users()->attach($user->id,['role' => 'project_owner']);
 
         $project->load(['users']);
 
@@ -54,11 +59,16 @@ class ProjectController extends ApiController
      */
     public function show(Request $request, Project $project): ProjectResource|JsonResponse
     {
-        // 自分がプロジェクトのメンバーかチェック（users()リレーションを使用）
+        // 自分がプロジェクトのメンバーかチェック（users()リレーションを使用）【Lesson2】
+        $user = $request->user();
+        $projectUser = $project->users()->where('users.id', $user->id)->exists();
 
-        // メンバーでなければエラーを返す
+        // メンバーでなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: このプロジェクトにアクセスする権限がありません
-
+        if (!$projectUser){
+            $message = ['message' => 'このプロジェクトにアクセスする権限がありません'];
+            return response()->json($message, 403);
+        }
 
         // 読み込み（N+1問題を防ぐため）
         $project->load(['users', 'tasks.createdBy']);
@@ -71,13 +81,25 @@ class ProjectController extends ApiController
      */
     public function update(Request $request, Project $project): ProjectResource|JsonResponse
     {
-        // 自分がオーナーまたは管理者かチェック（users()リレーションを使用）
+        // 自分がオーナーまたは管理者かチェック（users()リレーションを使用）【Lesson2】
+        $user = $request->user();
+        $ownerUser = $project->users()
+            ->where('users.id', $user->id)
+            ->wherePivotIn('role', ['project_owner', 'project_admin'])
+            ->exists();
 
-
-        // 権限がなければエラーを返す
+        // 権限がなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: プロジェクトを更新する権限がありません
+        if (!$ownerUser){
+            $message = ['message' => 'プロジェクトを更新する権限がありません'];
+            return response()->json($message, 403);
+        }
 
-        // バリデーションを記載
+        // バリデーションを記載【Lesson2】
+        $validated = $request->validate([
+            'name' => 'required',
+            'is_archived' => 'required',
+        ]);
 
         // プロジェクトを更新
         $project->update($request->only(['name', 'is_archived']));
@@ -94,10 +116,19 @@ class ProjectController extends ApiController
      */
     public function destroy(Request $request, Project $project): JsonResponse
     {
-        // 自分がオーナーかチェック（users()リレーションを使用
+        // 自分がオーナーかチェック（users()リレーションを使用【Lesson2】
+        $user = $request->user();
+        $ownerUser = $project->users()
+            ->where('users.id', $user->id)
+            ->wherePivotIn('role', ['project_owner'])
+            ->exists();
 
         // 権限がなければエラーを返す
-        // エラーコード: 403, エラーメッセージ: プロジェクトを削除する権限がありません
+        // エラーコード: 403, エラーメッセージ: プロジェクトを削除する権限がありません【Lesson2】
+        if (!$ownerUser){
+            $message = ['message' => 'プロジェクトを更新する権限がありません'];
+            return response()->json($message, 403);
+        }
 
         $project->delete();
