+++ app/Http/Controllers/Api/TaskController.php
@@ -17,11 +17,16 @@ class TaskController extends ApiController
      */
     public function index(Request $request, Project $project): AnonymousResourceCollection|JsonResponse
     {
-        // 自分がプロジェクトのメンバーかチェック
+        // 自分がプロジェクトのメンバーかチェック【Lesson2】
+        $user = $request->user();
+        $projectUser = $project->users()->where('users.id', $user->id)->exists();
 
-
-        // メンバーでなければエラーを返す
+        // メンバーでなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: このプロジェクトにアクセスする権限がありません
+        if (!$projectUser){
+            $message = ['message' => 'このプロジェクトにアクセスする権限がありません'];
+            return response()->json($message, 403);
+        }
 
         $tasks = $project->tasks()
             ->with('createdBy')
@@ -36,15 +41,23 @@ class TaskController extends ApiController
      */
     public function store(Request $request, Project $project): TaskResource|JsonResponse
     {
-        // 自分がプロジェクトのメンバーかチェック
+        // 自分がプロジェクトのメンバーかチェック【Lesson2】
+        $user = $request->user();
+        $projectUser = $project->users()->where('users.id', $user->id)->exists();
 
-
-        // メンバーでなければエラーを返す
+        // メンバーでなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: このプロジェクトにアクセスする権限がありません
+        if (!$projectUser){
+            $message = ['message' => 'このプロジェクトにアクセスする権限がありません'];
+            return response()->json($message, 403);
+        }
 
-
-        // バリデーションを記載
-
+        // バリデーションを記載【Lesson2】
+        $validated = $request->validate([
+            'title' => 'required',
+            'description' => 'nullable',
+            'status' => 'nullable',
+        ]);
 
         $task = Task::create([
             'project_id' => $project->id,
@@ -68,10 +81,17 @@ class TaskController extends ApiController
      */
     public function show(Request $request, Task $task): TaskResource|JsonResponse
     {
-        // 自分がプロジェクトのメンバーかチェック
-
-        // メンバーでなければエラーを返す
+        // 自分がプロジェクトのメンバーかチェック【Lesson2】
+        $user = $request->user();
+        $project = $task->project;
+        $projectUser = $project->users()->where('users.id', $user->id)->exists();
+        
+        // メンバーでなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: このプロジェクトにアクセスする権限がありません
+        if (!$projectUser){
+            $message = ['message' => 'このプロジェクトにアクセスする権限がありません'];
+            return response()->json($message, 403);
+        }
 
         // タスクを読み込み（N+1問題を防ぐため）
         $task->load(['createdBy', 'project']);
@@ -84,16 +104,28 @@ class TaskController extends ApiController
      */
     public function update(Request $request, Task $task): TaskResource|JsonResponse
     {
-        // 自分がプロジェクトのメンバーかチェック
-
-
-        // メンバーでなければエラーを返す
+        // 自分がプロジェクトのメンバーかチェック【Lesson2】
+        $user = $request->user();
+        $project = $task->project;
+        $projectUser = $project->users()->where('users.id', $user->id)->exists();
+        
+        // メンバーでなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: このプロジェクトにアクセスする権限がありません
+        if (!$projectUser){
+            $message = ['message' => 'このプロジェクトにアクセスする権限がありません'];
+            return response()->json($message, 403);
+        }
 
         // statusが“doneは更新不可（409）”を追加
         // エラーコード: 409, エラーメッジ: doneは更新不可です
 
-        // バリデーションを記載
+        // バリデーションを記載【Lesson2】
+        $validated = $request->validate([
+            'title' => 'required',
+            'description' => 'nullable',
+            'status' => 'nullable',
+        ]);
+
 
 
         $task->update($request->only(['title', 'description', 'status']));
@@ -110,12 +142,17 @@ class TaskController extends ApiController
      */
     public function destroy(Request $request, Task $task): JsonResponse
     {
-        // 自分がプロジェクトのメンバーかチェック
-
-
-        // メンバーでなければエラーを返す
+        // 自分がプロジェクトのメンバーかチェック【Lesson2】
+        $user = $request->user();
+        $project = $task->project;
+        $projectUser = $project->users()->where('users.id', $user->id)->exists();
+        
+        // メンバーでなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: このプロジェクトにアクセスする権限がありません
-
+        if (!$projectUser){
+            $message = ['message' => 'このプロジェクトにアクセスする権限がありません'];
+            return response()->json($message, 403);
+        }
 
         $task->delete();
 
@@ -129,13 +166,25 @@ class TaskController extends ApiController
      */
     public function start(Request $request, Task $task): TaskResource|JsonResponse
     {
-        // 自分がプロジェクトのメンバーかチェック
-
-        // メンバーでなければエラーを返す
+        // 自分がプロジェクトのメンバーかチェック【Lesson2】
+        $user = $request->user();
+        $project = $task->project;
+        $projectUser = $project->users()->where('users.id', $user->id)->exists();
+        
+        // メンバーでなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: このプロジェクトにアクセスする権限がありません
+        if (!$projectUser){
+            $message = ['message' => 'このプロジェクトにアクセスする権限がありません'];
+            return response()->json($message, 403);
+        };
 
         // ステータスがtodoでなければエラーを返す
         // エラーコード: 409, エラーメッセージ: 未着手のタスクのみ開始できます
+        $status= $task->status;
+        if ($status !== 'todo'){
+            $message = ['message' => '未着手のタスクのみ開始できます'];
+            return response()->json($message, 409);
+        }
 
         // タスクを開始
         $task->update(['status' => 'doing']);
@@ -152,13 +201,25 @@ class TaskController extends ApiController
      */
     public function complete(Request $request, Task $task): TaskResource|JsonResponse
     {
-        // 自分がプロジェクトのメンバーかチェック
-
-        // メンバーでなければエラーを返す
+        // 自分がプロジェクトのメンバーかチェック【Lesson2】
+        $user = $request->user();
+        $project = $task->project;
+        $projectUser = $project->users()->where('users.id', $user->id)->exists();
+        
+        // メンバーでなければエラーを返す【Lesson2】
         // エラーコード: 403, エラーメッセージ: このプロジェクトにアクセスする権限がありません
+        if (!$projectUser){
+            $message = ['message' => 'このプロジェクトにアクセスする権限がありません'];
+            return response()->json($message, 403);
+        }
 
         // ステータスがdoingでなければエラーを返す
         // エラーコード: 409, エラーメッセージ: 作業中のタスクのみ完了できます
+        $status= $task->status;
+        if ($status !== 'doing'){
+            $message = ['message' => '作業中のタスクのみ完了できます'];
+            return response()->json($message, 409);
+        }
 
         // タスクを完了
         $task->update(['status' => 'done']);
